separated generator from app
