+++ src/App.js
@@ -1,37 +1,30 @@
-import React from "react";
-import axios from "axios";
-import Joke from "./joke";
+import React from 'react';
+import logo from './logo.svg';
+import JokeGenerator from './JokeGenerator'
+import './App.css';
 
-const RANDOM_JOKE_URL = "https://api.icndb.com/jokes/random";
-
-export default class JokeGenerator extends React.Component {
-  state = {
-    joke: null,
-    loading: false
-  };
-
-  loadJoke = async () => {
-    this.setState({ loading: true });
-
-    const { data: { value: { joke } } } = await axios.get(RANDOM_JOKE_URL);
-
-    this.setState({ loading: false, joke });
-  };
-
-  render() {
-    const { joke, loading } = this.state;
-
-    return (
-      <React.Fragment>
-        {!joke && !loading && <div>You haven't loaded any joke yet!</div>}
-        {loading && <div>Loading...</div>}
-        {joke && !loading && <Joke text={joke} />}
-
-        <button onClick={this.loadJoke} type="button">
-          Load a random joke
-        </button>
-      </React.Fragment>
-    );
-  }
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <p> 
+          kasra jannati : 101237299
+          </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
+  );
 }
 
+export default App;
